Initial comit
Simplest version possible of a ressources table. It looks for files in the "database" folder, and if these files are resources, it loads it in a table.
The basic features are the following :
- displaying the resources
- possibility to double click and edit values of variables for each resources
- possibility to resize the rows

There are some bugs with the display of the table wich will be displayed on top of the left panel.